src/main/java/vnavesnoj/spring/validation/impl/UniqueUsernameConstraintValidator.java
package vnavesnoj.spring.validation.impl;

import jakarta.validation.ConstraintValidator;
import jakarta.validation.ConstraintValidatorContext;
import lombok.RequiredArgsConstructor;
import org.springframework.stereotype.Component;
import vnavesnoj.spring.database.repository.UserRepository;
import vnavesnoj.spring.validation.UniqueUsername;

/**
 * @author vnavesnoj
 * @mail vnavesnoj@gmail.com
 */
@Component
@RequiredArgsConstructor
public class UniqueUsernameConstraintValidator implements ConstraintValidator<UniqueUsername, String> {

    private final UserRepository userRepository;

    @Override
    public boolean isValid(String username, ConstraintValidatorContext context) {
        return username == null || !userRepository.existsByUsername(username.trim());
    }
}
